// frontend/src/types/schedule.ts
export type EmployeeTimeOff = {
    id: number;
    employeeId: number;
    employeeName: string;
    startDate: Date;
    endDate: Date;
    type: TimeOffType;
}

export enum TimeOffType {
    Vacation = "Vacation",
    SickLeave = "SickLeave",
    PersonalDay = "PersonalDay",
}

export type Shift = {
    id: number;
    date: string;
    shiftTypeId: number;
    shiftTypeName: string;
    startTime: string;
    endTime: string;
    color: string;
    employeeNeeded: number;
    employees: EmployeeShiftAssignment[];
}

export type EmployeeShiftAssignment = {
    id: number;
    name: string;
    departmentName?: string;
    note?: string;
}